+++ Locales/enUS.lua
@@ -87,6 +87,22 @@ select(2, ...).L = setmetatable({
         <p>(Retail) Enable "Always Update Buffs/Debuffs" in General tab, if indicators do not update correctly.</p>
         <br/>
 
+        <h1>r237-release (Aug 6, 2024, 21:30 GMT+8)</h1>
+        <p>* (TWW) Updated default indicator spells (PR #165).</p>
+        <p>* Updated gradient colors (PR #181).</p>
+        <p>+ New snippet var CELL_RANGE_CHECK_*. Custom spells can now be used for range checking.</p>
+        <p>+ New custom indicator type: Bars.</p>
+        <p>+ Added "Health Bar (Loss)" option to Color indicator.</p>
+        <p>+ (TWW) Added Skyfury to Buff Tracker.</p>
+        <p>* Updated layer of Health Thresholds indicator.</p>
+        <p>* Slightly optimized Cell.GetUnitFramesForLGF.</p>
+        <p>* Fixed an indicator loading issue.</p>
+        <p>* Fixed some click-casting issues.</p>
+        <p>* (TWW) Fixed dispel check.</p>
+        <p>* (Classic) Fixed Spell Request.</p>
+        <p>* Updated deDE and zhTW.</p>
+        <br/>
+
         <h1>r236-release (Jul 24, 2024, 16:10 GMT+8)</h1>
         <p>* Fixed appearance/layout tab.</p>
         <p>* Updated locales.</p>